Add ames, dili and carcinogens docs for TDC (#412)
