+++ docs/modules/datasets.rst
@@ -9,3 +9,4 @@ Functions for loading benchmark molecular datasets.
 
     datasets/lrgb
     datasets/moleculenet
+    datasets/tdc
