+++ docs/modules/datasets/tdc.rst
@@ -1,0 +1,29 @@
+=================================
+TDC
+=================================
+
+.. automodule:: skfp.datasets.tdc
+    :exclude-members: load_tdc_benchmark, load_tdc_splits
+
+=========================================================
+
+.. py:currentmodule:: skfp.datasets.tdc
+
+TDC molecular datasets
+
+.. autosummary::
+    :nosignatures:
+    :toctree: generated/
+
+    load_tdc_benchmark
+    load_tdc_splits
+
+Dataset loaders
+
+.. autosummary::
+    :nosignatures:
+    :toctree: generated/
+
+    tox.load_ames
+    tox.load_dili
+    tox.load_carcinogens_lagunin
