+++ tests/datasets/tdc.py
@@ -35,16 +35,18 @@ from skfp.datasets.tdc.adme.solubility_aqsoldb import load_solubility_aqsoldb
 from skfp.datasets.tdc.adme.vdss_lombardo import load_vdss_lombardo
 from skfp.datasets.tdc.hts.sarscov2_3clpro_diamond import load_sarscov2_3clpro_diamond
 from skfp.datasets.tdc.hts.sarscov2_vitro_touret import load_sarscov2_vitro_touret
-from skfp.datasets.tdc.tox.ames import load_ames
-from skfp.datasets.tdc.tox.carcinogens_lagunin import load_carcinogens_lagunin
-from skfp.datasets.tdc.tox.dili import load_dili
-from skfp.datasets.tdc.tox.herg import load_herg
-from skfp.datasets.tdc.tox.herg_central_at_1um import load_herg_central_at_1um
-from skfp.datasets.tdc.tox.herg_central_at_10um import load_herg_central_at_10um
-from skfp.datasets.tdc.tox.herg_central_inhib import load_herg_central_inhib
-from skfp.datasets.tdc.tox.herg_karim import load_herg_karim
-from skfp.datasets.tdc.tox.ld50_zhu import load_ld50_zhu
-from skfp.datasets.tdc.tox.skin_reaction import load_skin_reaction
+from skfp.datasets.tdc.tox import (
+    load_ames,
+    load_carcinogens_lagunin,
+    load_dili,
+    load_herg,
+    load_herg_central_at_1um,
+    load_herg_central_at_10um,
+    load_herg_central_inhib,
+    load_herg_karim,
+    load_ld50_zhu,
+    load_skin_reaction,
+)
 from tests.datasets.test_utils import run_basic_dataset_checks
